Rearrange a bunch of things to implement e-stop button from Mission Control
