+++ src/Globals.h
@@ -21,4 +21,6 @@ namespace Globals
     extern RoverState curr_state;
     extern nlohmann::json status_data;
     extern nlohmann::json motor_status;
+    extern bool E_STOP;
+    extern bool AUTONOMOUS;
 }
