+++ src/Networking/NetworkConstants.h
@@ -11,7 +11,7 @@
 #include <cstring>
 #include <cstdlib>
 
-#define PORT 5000 
+#define PORT 3001
 #define MAXLINE 1024
 #define CLOSE_TCP "~~closetcp~~"
 #define SERVER_IP "127.0.0.1"
