+++ src/Networking/ParseBaseStation.h
@@ -5,7 +5,6 @@
 #include "json.hpp"
 using nlohmann::json;
 bool ParseBaseStationPacket(char const* buffer);
-bool ParseMotorPacket(json &message);
 bool sendError(std::string const &msg);
 
 #endif
