+++ src/Networking/motor_interface.h
@@ -1,0 +1,7 @@
+
+#pragma once
+
+#include "json.hpp"
+using nlohmann::json;
+
+bool ParseMotorPacket(json &message);
